Remove std::io once and for all!
