+++ src/test/compile-fail/borrowck-auto-mut-ref-to-immut-var.rs
@@ -10,8 +10,6 @@
 
 // Tests that auto-ref can't create mutable aliases to immutable memory.
 
-use std::io;
-
 struct Foo {
     x: int
 }
